pickup/exp tree done

+++ Assets/Scripts/Prefs/SendToAufburn.cs
@@ -25,7 +25,7 @@ public class SendToAufburn : MonoBehaviour
     public IEnumerator SendBackToAufburn()
     {
         
-        PlayerPrefs.SetInt("Gold", playerCombat.playerMoney);
+        PlayerPrefs.SetFloat("Gold", playerCombat.playerMoney);
         PlayerPrefs.SetInt("Exp", playerCombat.playerXp);
         PlayerPrefs.SetInt("DelveLevel", playerCombat.delveLevel);
         PlayerPrefs.Save();
